+++ ts/EventName.ts
@@ -1,0 +1,11 @@
+'use strict'
+
+enum EventName {
+    handleUrlCalled = "handleUrlCalled",
+    handleCustomActionCalled = "handleCustomActionCalled",
+    handleInAppCalled = "handleInAppCalled",
+    handleAuthCalled = "handleAuthCalled",
+    receivedIterableInboxChanged = "receivedIterableInboxChanged"
+}
+
+export default EventName
